requirements.txt
yfinance
pandas
numpy
matplotlib
scikit-learn
newsapi-python
transformers
torch
